move web package into core directory
