+++ core/web/bulk_deletes_controller.go
