+++ core/web/gui_assets_test.go
